Rework Enchant Weapon

+++ code/modules/spells/components/enchanted_weapon.dm
@@ -1,53 +1,50 @@
-#define DEFAULT_DURATION 15 MINUTES
+#define DEFAULT_DURATION 10 MINUTES
+#define ENCHANT_REFRESH_INTERVAL 1 MINUTES
+#define ENCHANT_PROXIMITY_RANGE 10
 
 // I nuked the former Searing Blade. Leaving 1 open for a future replacement
 
 /* Component used for adding enchantment from the enchant weapon spell
- Three types of enchantments are available:
+ Two types of enchantments are available:
  1. Force Blade: Increases the force of the weapon by 5.
  2. Durability: Increases the integrity and max integrity of the weapon by 100.
- 3. Arcane Mark: Applies a stack of Arcane Mark, 7 sec cd (ARCANE_MARK_COOLDOWN)
- The enchantment will lasts for 15 minutes, and will automatically refresh in the hand of an Arcyne user.
+ The enchantment lasts for 10 minutes, and will automatically refresh while within 10 tiles of the original caster.
+ A beam is drawn between caster and holder on refresh (same z-level only).
  There used to be a concept for a blade to set people on fire - but it was too broken if people didn't insta pat
 */
 
 /datum/component/enchanted_weapon
 	dupe_mode = COMPONENT_DUPE_UNIQUE // To avoid weird filter override this is the way..
 	var/endtime = null // How long till the conjured item disappear (Don't use duration it is messed up)
-	var/allow_refresh = TRUE // If TRUE, the item will refresh its duration when held by a X user
-	var/refresh_skill = /datum/skill/magic/arcane // The skill that will be used to refresh the item
+	var/allow_refresh = TRUE // If TRUE, the item will refresh its duration when within range of the caster
 	var/overridden_duration = null
 	var/enchant_type = FORCE_BLADE_ENCHANT // The type of enchantment
-	var/next_arcane_mark_time = 0
+	var/datum/weakref/caster_ref = null // Weakref to the original caster for proximity refresh
 	var/datum/weakref/last_known_mob = null // Cached mob weakref for when weapon is in nullspace (e.g. holster)
 
-/datum/component/enchanted_weapon/Initialize(duration_override, allow_refresh_override, refresh_skill_override, enchant_type_override)
+/datum/component/enchanted_weapon/Initialize(duration_override, allow_refresh_override, caster, enchant_type_override)
 	if(!istype(parent, /obj/item/rogueweapon))
 		return COMPONENT_INCOMPATIBLE
 	var/obj/item/I = parent
-
-	var/new_duration = duration_override ? duration_override : DEFAULT_DURATION
 
 	if(duration_override)
 		endtime = world.time + duration_override
 		overridden_duration = duration_override
 	else
 		endtime = world.time + DEFAULT_DURATION
-	if(allow_refresh_override)
+	if(!isnull(allow_refresh_override))
 		allow_refresh = allow_refresh_override
-	if(refresh_skill_override)
-		refresh_skill = refresh_skill_override
 	if(enchant_type_override)
 		enchant_type = enchant_type_override
+	if(caster)
+		caster_ref = WEAKREF(caster)
 	apply_enchant(I)
 
 	RegisterSignal(parent, COMSIG_PARENT_EXAMINE, PROC_REF(on_examine))
 	RegisterSignal(parent, COMSIG_ITEM_OBJFIX, PROC_REF(on_fix))
 	RegisterSignal(parent, COMSIG_MOVABLE_MOVED, PROC_REF(on_moved))
-	if(enchant_type == ARCANE_MARK_ENCHANT)
-		RegisterSignal(parent, COMSIG_ITEM_AFTERATTACK, PROC_REF(arcane_mark_afterattack))
 
-	addtimer(CALLBACK(src, PROC_REF(refresh_check)), new_duration)
+	addtimer(CALLBACK(src, PROC_REF(refresh_check)), ENCHANT_REFRESH_INTERVAL)
 
 /datum/component/enchanted_weapon/proc/on_moved(datum/source, atom/old_loc, dir, forced)
 	SIGNAL_HANDLER
@@ -60,42 +57,49 @@
 		check = check.loc
 	last_known_mob = null
 
-/datum/component/enchanted_weapon/proc/refresh_check()
+/datum/component/enchanted_weapon/proc/get_holder_mob()
 	var/obj/item/I = parent
 	var/atom/itemloc = I.loc
-	if(!allow_refresh || !refresh_skill)
-		remove()
-		qdel(src)
-		return
-	// If loc is null, weapon is in nullspace (e.g. holster component). Use cached mob.
 	if(isnull(itemloc))
 		var/mob/living/cached_mob = last_known_mob?.resolve()
 		if(cached_mob)
-			itemloc = cached_mob
-		else
-			qdel(src)
-			return
-	if(!istype(itemloc, /mob/living))
-		while(!istype(itemloc, /mob/living))
-			if(isnull(itemloc))
-				qdel(src)
-				return // If the item is not in a mob, remove the refresh
-			itemloc = itemloc.loc
-			if(istype(itemloc, /turf))
-				qdel(src)
-				return // If the item is on the ground, remove the refresh
-	var/mob/living/current_user = itemloc
-	var/has_right_skill = current_user?.get_skill_level(refresh_skill)
-	if(has_right_skill)
-		if(overridden_duration)
-			endtime = world.time + overridden_duration
-			addtimer(CALLBACK(src, PROC_REF(refresh_check)), overridden_duration)
-		else
-			endtime = world.time + DEFAULT_DURATION
-			addtimer(CALLBACK(src, PROC_REF(refresh_check)), DEFAULT_DURATION)
-	else
+			return cached_mob
+		return null
+	while(itemloc && !isturf(itemloc))
+		if(isliving(itemloc))
+			return itemloc
+		itemloc = itemloc.loc
+	return null
+
+/datum/component/enchanted_weapon/proc/refresh_check()
+	if(!allow_refresh || !caster_ref)
 		qdel(src)
 		return
+	var/mob/living/caster = caster_ref.resolve()
+	if(!caster || caster.stat == DEAD)
+		qdel(src)
+		return
+	var/mob/living/holder = get_holder_mob()
+	if(!holder)
+		qdel(src)
+		return
+	var/turf/caster_turf = get_turf(caster)
+	var/turf/holder_turf = get_turf(holder)
+	if(!caster_turf || !holder_turf)
+		qdel(src)
+		return
+	if(get_dist(caster_turf, holder_turf) > ENCHANT_PROXIMITY_RANGE)
+		qdel(src)
+		return
+	// Refresh the duration
+	if(overridden_duration)
+		endtime = world.time + overridden_duration
+	else
+		endtime = world.time + DEFAULT_DURATION
+	// Draw beam if on the same z-level and holder is not the caster
+	if(holder != caster && caster_turf.z == holder_turf.z)
+		caster.Beam(holder, icon_state = "b_beam", time = 5, maxdistance = ENCHANT_PROXIMITY_RANGE)
+	addtimer(CALLBACK(src, PROC_REF(refresh_check)), ENCHANT_REFRESH_INTERVAL)
 
 /datum/component/enchanted_weapon/proc/apply_enchant(var/obj/item/I, is_fix = FALSE)
 	if(enchant_type == FORCE_BLADE_ENCHANT)
@@ -112,10 +116,6 @@
 		var/durability_filter = I.get_filter(DURABILITY_FILTER)
 		if(!durability_filter)
 			I.add_filter(DURABILITY_FILTER, 2, list("type" = "outline", "color" = GLOW_COLOR_METAL, "alpha" = 200, "size" = 1))
-	else if(enchant_type == ARCANE_MARK_ENCHANT)
-		var/mark_filter = I.get_filter(ARCANE_MARK_FILTER_WEAPON)
-		if(!mark_filter)
-			I.add_filter(ARCANE_MARK_FILTER_WEAPON, 2, list("type" = "outline", "color" = GLOW_COLOR_ARCANE, "alpha" = 200, "size" = 1))
 
 // Called when the enchantment is removed
 /datum/component/enchanted_weapon/proc/remove()
@@ -129,13 +129,12 @@
 			I.max_integrity -= DURABILITY_INCREASE // Jank ass "temporary" fix I sure hope no one else modify max integrity
 		I.obj_integrity = min(I.obj_integrity, I.max_integrity - DURABILITY_INCREASE)
 		I.remove_filter(DURABILITY_FILTER)
-	else if(enchant_type == (ARCANE_MARK_ENCHANT))
-		I.remove_filter(ARCANE_MARK_FILTER_WEAPON)
 	else
 		return
 
 /datum/component/enchanted_weapon/Destroy()
 	remove()
+	caster_ref = null
 	last_known_mob = null
 	. = ..()
 
@@ -144,8 +143,6 @@
 		examine_list += "This weapon is enchanted with a force blade enchantment."
 	else if(enchant_type == DURABILITY_ENCHANT)
 		examine_list += "This weapon is enchanted with a durability enchantment."
-	else if(enchant_type == ARCANE_MARK_ENCHANT)
-		examine_list += "This weapon is enchanted with an Arcane Mark enchantment."
 	var/remaining_minutes = round((endtime - world.time) / 600)
 	examine_list += "The enchantment will last for [remaining_minutes] more minutes."
 
@@ -154,20 +151,6 @@
 	var/obj/item/I = parent
 	apply_enchant(I, TRUE)
 
-//special snowflake arcane mark proc grahhh
-/datum/component/enchanted_weapon/proc/arcane_mark_afterattack(obj/item/source, atom/target, mob/user, proximity_flag, click_parameters)
-
-	SIGNAL_HANDLER
-	if(!proximity_flag)
-		return
-	if(world.time < next_arcane_mark_time)
-		return
-	if(!ismob(target))
-		return
-	var/mob/living/L = target
-	if(istype(L, /mob/living/carbon))
-		apply_arcane_mark(L)
-		playsound(get_turf(L), 'sound/magic/clang.ogg', 100)
-		next_arcane_mark_time = world.time + ARCANE_MARK_COOLDOWN
-
 #undef DEFAULT_DURATION
+#undef ENCHANT_REFRESH_INTERVAL
+#undef ENCHANT_PROXIMITY_RANGE
